move all preloading into libsbdtestbench.so
