+++ Makefile.am
@@ -28,7 +28,7 @@ COUNT           = $(shell expr 1 + $(LAST_COUNT))
 
 TESTS           = tests/regressions.sh
 export SBD_BINARY := src/sbd
-export SBD_PRELOAD := tests/.libs/libfakeblock.so
+export SBD_PRELOAD := tests/.libs/libsbdtestbed.so
 export SBD_USE_DM := no
 
 EXTRA_DIST      = sbd.spec tests/regressions.sh
