+++ tests/Makefile.am
@@ -1,5 +1,3 @@
-lib_LTLIBRARIES = libfakeaio.la libfakeblock.la
-libfakeaio_la_SOURCES = sbd-fakeaio.c
-libfakeaio_la_LDFLAGS =
-libfakeblock_la_SOURCES = sbd-fakeblock.c
-libfakeblock_la_LDFLAGS =
+lib_LTLIBRARIES = libsbdtestbed.la
+libsbdtestbed_la_SOURCES = sbd-testbed.c
+libsbdtestbed_la_LDFLAGS =
