+++ tests/sbd-testbed.c
@@ -9,6 +9,7 @@
 #include <stdarg.h>
 #include <stddef.h>
 #include <fcntl.h>
+#include <libaio.h>
 #include <linux/watchdog.h>
 #include <linux/fs.h>
 #include <stdio.h>
@@ -37,7 +38,7 @@ GIOFunc2GUnixFDSourceFunc(GIOChannel *source,
 }
 
 static guint
-g_unix_fd_add (gint fd,
+g_unix_fd_add(gint fd,
                GIOCondition condition,
                GUnixFDSourceFunc function,
                gpointer user_data)
@@ -61,6 +62,12 @@ typedef ssize_t (*orig_write_f_type)(int fd, const void *buf, size_t count);
 typedef int (*orig_close_f_type)(int fd);
 typedef FILE *(*orig_fopen_f_type)(const char *pathname, const char *mode);
 typedef int (*orig_fclose_f_type)(FILE *fp);
+typedef int (*orig_io_setup_f_type)(int nr_events, io_context_t *ctx_idp);
+typedef int (*orig_io_submit_f_type)(io_context_t ctx_id, long nr, struct iocb *ios[]);
+typedef int (*orig_io_getevents_f_type)(io_context_t ctx_id, long min_nr, long nr,
+                struct io_event *events, struct timespec *timeout);
+typedef int (*orig_io_cancel_f_type)(io_context_t ctx_id, struct iocb *iocb,
+                struct io_event *result);
 
 static int is_init = 0;
 static FILE *log_fp = NULL;
@@ -85,12 +92,23 @@ static orig_write_f_type orig_write = NULL;
 static orig_close_f_type orig_close = NULL;
 static orig_fopen_f_type orig_fopen = NULL;
 static orig_fclose_f_type orig_fclose = NULL;
+static orig_io_setup_f_type orig_io_setup = NULL;
+static orig_io_submit_f_type orig_io_submit = NULL;
+static orig_io_getevents_f_type orig_io_getevents = NULL;
+static orig_io_cancel_f_type orig_io_cancel = NULL;
 
-/* For fprintf is inlined as __fprintf_chk.
+/* fprintf is inlined as __fprintf_chk or
+ * we have vfprintf.
  * For fscanf we have vfscanf.
  * For reboot we anyway don't want that to be
  * called in any case.
  */
+
+static struct iocb *pending_iocb = NULL;
+struct io_context { int context_num; };
+static struct io_context our_io_context = {.context_num = 1};
+static int translate_aio = 0;
+
 static GMainLoop *mainloop = NULL;
 
 #if 0
@@ -102,6 +120,87 @@ watchdog_shutdown(int nsig)
     }
 }
 #endif
+
+static void*
+dlsym_fatal(void *handle, const char *symbol)
+{
+    void *rv = dlsym(handle, symbol);
+
+    if (!rv) {
+        fprintf(stderr, "Failed looking up symbol %s\n", symbol);
+        exit(1);
+    }
+    return rv;
+}
+
+static void
+init (void)
+{
+    void *handle;
+
+    if (!is_init) {
+        const char *value;
+        int i;
+        char *token, *str, *str_orig;
+
+        is_init = 1;
+
+        orig_open         = (orig_open_f_type)dlsym_fatal(RTLD_NEXT,"open");
+        orig_ioctl        = (orig_ioctl_f_type)dlsym_fatal(RTLD_NEXT,"ioctl");
+        orig_close        = (orig_close_f_type)dlsym_fatal(RTLD_NEXT,"close");
+        orig_write        = (orig_write_f_type)dlsym_fatal(RTLD_NEXT,"write");
+        orig_fopen        = (orig_fopen_f_type)dlsym_fatal(RTLD_NEXT,"fopen");
+        orig_fclose       = (orig_fclose_f_type)dlsym_fatal(RTLD_NEXT,"fclose");
+
+        handle = dlopen("libaio.so.1",  RTLD_NOW);
+        if (!handle) {
+            fprintf(stderr, "Failed opening libaio.so.1\n");
+            exit(1);
+        }
+        orig_io_setup     = (orig_io_setup_f_type)dlsym_fatal(handle,"io_setup");
+        orig_io_submit    = (orig_io_submit_f_type)dlsym_fatal(handle,"io_submit");
+        orig_io_getevents = (orig_io_getevents_f_type)dlsym_fatal(handle,"io_getevents");
+        orig_io_cancel    = (orig_io_cancel_f_type)dlsym_fatal(handle,"io_cancel");
+        dlclose(handle);
+
+        value = getenv("SBD_PRELOAD_LOG");
+        if (value) {
+            log_fp = fopen(value, "a");
+        } else {
+            int fd = dup(fileno(stderr));
+            if (fd >= 0) {
+                log_fp = fdopen(fd, "w");
+            }
+        }
+        if (log_fp == NULL) {
+            fprintf(stderr, "couldn't open log-file\n");
+        }
+
+        value = getenv("SBD_WATCHDOG_DEV");
+        if (value) {
+            watchdog_device = strdup(value);
+        }
+
+        value = getenv("SBD_DEVICE");
+        if ((value) && (str = str_orig = strdup(value))) {
+            for (i = 0; i < 3; i++, str = NULL) {
+                token = strtok(str, ";");
+                if (token == NULL) {
+                    break;
+                }
+                sbd_device[i] = strdup(token);
+            }
+            free(str_orig);
+        }
+
+        value = getenv("SBD_TRANSLATE_AIO");
+        if ((value) && !strcmp(value, "yes")) {
+            translate_aio = 1;
+        }
+    }
+}
+
+// ***** end - handling of watchdog & block-devices ****
 
 static gboolean
 watchdog_timeout_notify(gpointer data)
@@ -152,54 +251,6 @@ watchdog_dispatch_callback (gint fd,
         fprintf(log_fp, "unknown watchdog command\n");
     }
     return TRUE;
-}
-
-static void
-init (void)
-{
-    if (!is_init) {
-        const char *value;
-        int i;
-        char *token, *str, *str_orig;
-
-        is_init = 1;
-
-        orig_open    = (orig_open_f_type)dlsym(RTLD_NEXT,"open");
-        orig_ioctl   = (orig_ioctl_f_type)dlsym(RTLD_NEXT,"ioctl");
-        orig_close   = (orig_close_f_type)dlsym(RTLD_NEXT,"close");
-        orig_write   = (orig_write_f_type)dlsym(RTLD_NEXT,"write");
-        orig_fopen   = (orig_fopen_f_type)dlsym(RTLD_NEXT,"fopen");
-        orig_fclose  = (orig_fclose_f_type)dlsym(RTLD_NEXT,"fclose");
-
-        value = getenv("SBD_PRELOAD_LOG");
-        if (value) {
-            log_fp = fopen(value, "a");
-        } else {
-            int fd = dup(fileno(stderr));
-            if (fd >= 0) {
-                log_fp = fdopen(fd, "w");
-            }
-        }
-        if (log_fp == NULL) {
-            fprintf(stderr, "couldn't open log-file\n");
-        }
-
-        value = getenv("SBD_WATCHDOG_DEV");
-        if (value) {
-            watchdog_device = strdup(value);
-        }
-        value = getenv("SBD_DEVICE");
-        if ((value) && (str = str_orig = strdup(value))) {
-            for (i = 0; i < 3; i++, str = NULL) {
-                token = strtok(str, ";");
-                if (token == NULL) {
-                    break;
-                }
-                sbd_device[i] = strdup(token);
-            }
-            free(str_orig);
-        }
-    }
 }
 
 static void
@@ -313,46 +364,6 @@ open(const char *pathname, int flags, ...)
     return fd;
 }
 
-FILE *
-fopen(const char *pathname, const char *mode)
-{
-    int is_sysrq = 0;
-    int is_sysrq_trigger = 0;
-    FILE *fp;
-
-    init();
-
-    if ((strcmp("/proc/sys/kernel/sysrq", pathname) == 0) &&
-        strcmp("w", mode)) {
-        pathname = "/dev/null";
-        is_sysrq = 1;
-    } else if (strcmp("/proc/sysrq-trigger", pathname) == 0) {
-        pathname = "/dev/null";
-        is_sysrq_trigger = 1;
-    }
-    fp = orig_fopen(pathname, mode);
-    if (is_sysrq) {
-        sysrq_fp = fp;
-    } else if (is_sysrq_trigger) {
-        sysrq_trigger_fp = fp;
-    }
-    return fp;
-}
-
-int
-fclose(FILE *fp)
-{
-    init();
-
-    if (fp == sysrq_fp) {
-        sysrq_fp = NULL;
-    } else if (fp == sysrq_trigger_fp) {
-        sysrq_trigger_fp = NULL;
-    }
-    return orig_fclose(fp);
-}
-
-
 ssize_t
 write(int fd, const void *buf, size_t count)
 {
@@ -446,6 +457,49 @@ close(int fd)
     return orig_close(fd);
 }
 
+// ***** end - handling of watchdog & block-devices ****
+
+// ***** handling of sysrq, sysrq-trigger & reboot ****
+
+FILE *
+fopen(const char *pathname, const char *mode)
+{
+    int is_sysrq = 0;
+    int is_sysrq_trigger = 0;
+    FILE *fp;
+
+    init();
+
+    if ((strcmp("/proc/sys/kernel/sysrq", pathname) == 0) &&
+        strcmp("w", mode)) {
+        pathname = "/dev/null";
+        is_sysrq = 1;
+    } else if (strcmp("/proc/sysrq-trigger", pathname) == 0) {
+        pathname = "/dev/null";
+        is_sysrq_trigger = 1;
+    }
+    fp = orig_fopen(pathname, mode);
+    if (is_sysrq) {
+        sysrq_fp = fp;
+    } else if (is_sysrq_trigger) {
+        sysrq_trigger_fp = fp;
+    }
+    return fp;
+}
+
+int
+fclose(FILE *fp)
+{
+    init();
+
+    if (fp == sysrq_fp) {
+        sysrq_fp = NULL;
+    } else if (fp == sysrq_trigger_fp) {
+        sysrq_trigger_fp = NULL;
+    }
+    return orig_fclose(fp);
+}
+
 #if __USE_FORTIFY_LEVEL > 1
 
 int
@@ -530,3 +584,163 @@ reboot (int __howto)
     killpg(0, SIGKILL);
     exit(1);
 }
+
+// ***** end - handling of sysrq, sysrq-trigger & reboot ****
+
+// ***** aio translate ****
+
+#if 0
+struct iocb {
+    void *data;
+    unsigned key;
+    short aio_lio_opcode;
+    short aio_reqprio;
+    int aio_fildes;
+};
+
+static inline void io_prep_pread(struct iocb *iocb, int fd, void *buf, size_t count, long long offset)
+{
+	memset(iocb, 0, sizeof(*iocb));
+	iocb->aio_fildes = fd;
+	iocb->aio_lio_opcode = IO_CMD_PREAD;
+	iocb->aio_reqprio = 0;
+	iocb->u.c.buf = buf;
+	iocb->u.c.nbytes = count;
+	iocb->u.c.offset = offset;
+}
+
+static inline void io_prep_pwrite(struct iocb *iocb, int fd, void *buf, size_t count, long long offset)
+{
+	memset(iocb, 0, sizeof(*iocb));
+	iocb->aio_fildes = fd;
+	iocb->aio_lio_opcode = IO_CMD_PWRITE;
+	iocb->aio_reqprio = 0;
+	iocb->u.c.buf = buf;
+	iocb->u.c.nbytes = count;
+	iocb->u.c.offset = offset;
+}
+#endif
+
+int io_setup(int nr_events, io_context_t *ctx_idp)
+{
+    init();
+
+    if (!translate_aio) {
+        return orig_io_setup(nr_events, ctx_idp);
+    }
+
+    if (nr_events == 0) {
+        return EINVAL;
+    }
+    if (nr_events > 1) {
+        return EAGAIN;
+    }
+    if (ctx_idp == NULL) {
+        return EFAULT;
+    }
+    *ctx_idp = &our_io_context;
+    return 0;
+}
+
+
+int io_submit(io_context_t ctx_id, long nr, struct iocb *ios[])
+{
+    init();
+
+    if (!translate_aio) {
+        return orig_io_submit(ctx_id, nr, ios);
+    }
+
+    if ((pending_iocb != NULL) ||
+        (nr > 1)) {
+        return EAGAIN;
+    }
+    if ((nr == 1) && ((ios == NULL) || (ios[0] == NULL))) {
+        return EFAULT;
+    }
+    if ((ctx_id != &our_io_context) ||
+        (nr < 0) ||
+        ((nr == 1) &&
+         (ios[0]->aio_lio_opcode != IO_CMD_PREAD) &&
+         (ios[0]->aio_lio_opcode != IO_CMD_PWRITE))) {
+        return EINVAL;
+    }
+    if ((fcntl(ios[0]->aio_fildes, F_GETFD) == -1) && (errno == EBADF)) {
+        return EBADF;
+    }
+    if (nr == 1) {
+        pending_iocb = ios[0];
+    }
+    return nr;
+}
+
+int io_getevents(io_context_t ctx_id, long min_nr, long nr,
+                        struct io_event *events, struct timespec *timeout)
+{
+    init();
+
+    if (!translate_aio) {
+        return orig_io_getevents(ctx_id, min_nr, nr, events, timeout);
+    }
+
+    if ((ctx_id != &our_io_context) ||
+        (min_nr != 1) ||
+        (nr != 1)) {
+        return EINVAL;
+    }
+    if (pending_iocb == NULL) {
+        return 0;
+    }
+
+	switch (pending_iocb->aio_lio_opcode) {
+		case IO_CMD_PWRITE:
+			events->res = pwrite(pending_iocb->aio_fildes,
+								pending_iocb->u.c.buf,
+								pending_iocb->u.c.nbytes,
+								pending_iocb->u.c.offset);
+			break;
+		case IO_CMD_PREAD:
+			events->res = pread(pending_iocb->aio_fildes,
+								pending_iocb->u.c.buf,
+								pending_iocb->u.c.nbytes,
+								pending_iocb->u.c.offset);
+			break;
+		default:
+			events->res = 0;
+	}
+
+    events->data = pending_iocb->data;
+    events->obj  = pending_iocb;
+
+    events->res2  = 0;
+    pending_iocb = NULL;
+    return 1;
+}
+
+int io_cancel(io_context_t ctx_id, struct iocb *iocb,
+                     struct io_event *result)
+{
+    init();
+
+    if (!translate_aio) {
+        return orig_io_cancel(ctx_id, iocb, result);
+    }
+
+    if (ctx_id != &our_io_context) {
+        return EINVAL;
+    }
+    if ((iocb == NULL) || (result == NULL)) {
+        return EFAULT;
+    }
+    if (pending_iocb != iocb) {
+        return EAGAIN;
+    }
+    result->data = iocb->data;
+    result->obj  = iocb;
+    result->res  = 0;
+    result->res2  = 0;
+    pending_iocb = NULL;
+    return 0;
+}
+
+// ***** end - aio translate ****
